Move Composed to Typeclasses
